+++ src/Admin.php
@@ -4,7 +4,6 @@ namespace SleepingOwl\Admin;
 
 use Closure;
 use Illuminate\Filesystem\Filesystem;
-use SleepingOwl\Admin\Navigation\Page;
 use Illuminate\Contracts\Support\Renderable;
 use SleepingOwl\Admin\Model\ModelCollection;
 use Illuminate\Foundation\ProviderRepository;
@@ -13,9 +12,12 @@ use SleepingOwl\Admin\Contracts\AdminInterface;
 use SleepingOwl\Admin\Model\ModelConfiguration;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Config\Repository as ConfigRepository;
+use SleepingOwl\Admin\Providers\AdminServiceProvider;
 use SleepingOwl\Admin\Contracts\Template\MetaInterface;
 use SleepingOwl\Admin\Http\Controllers\AdminController;
+use SleepingOwl\Admin\Providers\AliasesServiceProvider;
 use SleepingOwl\Admin\Contracts\Template\TemplateInterface;
+use SleepingOwl\Admin\Providers\BreadcrumbsServiceProvider;
 use SleepingOwl\Admin\Configuration\ProvidesScriptVariables;
 use SleepingOwl\Admin\Contracts\ModelConfigurationInterface;
 use SleepingOwl\Admin\Contracts\Navigation\NavigationInterface;
@@ -85,8 +87,8 @@ class Admin implements AdminInterface
     /**
      * @param string $class
      * @param Closure|null $callback
-     *
-     * @return $this
+     * @return $this|AdminInterface
+     * @throws Exceptions\RepositoryException
      */
     public function registerModel($class, Closure $callback = null)
     {
@@ -155,8 +157,9 @@ class Admin implements AdminInterface
     }
 
     /**
-     * @param string|Model $class
-     * @return ModelConfigurationInterface
+     * @param string $class
+     * @return mixed|null|ModelConfigurationInterface
+     * @throws Exceptions\RepositoryException
      */
     public function getModel($class)
     {
@@ -214,24 +217,14 @@ class Admin implements AdminInterface
     }
 
     /**
-     * @param string $class
-     * @param int    $priority
-     *
-     * @return Page
+     * @param null $class
+     * @param int $priority
+     * @return mixed
+     * @throws Exceptions\RepositoryException
      */
     public function addMenuPage($class = null, $priority = 100)
     {
         return $this->getModel($class)->addToNavigation($priority);
-    }
-
-    /**
-     * @return Navigation
-     *
-     * @deprecated
-     */
-    public function getNavigation()
-    {
-        return $this->navigation();
     }
 
     /**
@@ -253,10 +246,10 @@ class Admin implements AdminInterface
     protected function registerBaseServiceProviders()
     {
         $providers = [
-            \SleepingOwl\Admin\Providers\AliasesServiceProvider::class,
+            AliasesServiceProvider::class,
             \Collective\Html\HtmlServiceProvider::class,
-            \SleepingOwl\Admin\Providers\BreadcrumbsServiceProvider::class,
-            \SleepingOwl\Admin\Providers\AdminServiceProvider::class,
+            BreadcrumbsServiceProvider::class,
+            AdminServiceProvider::class,
         ];
 
         /* Workaround to allow use ServiceProvider-based configurations in old fashion */