+++ src/Display/DisplayDatatablesAsync.php
@@ -6,9 +6,6 @@ use Request;
 use Illuminate\Routing\Router;
 use Illuminate\Support\Collection;
 use Illuminate\Database\Eloquent\Builder;
-use SleepingOwl\Admin\Display\Column\Link;
-use SleepingOwl\Admin\Display\Column\Text;
-use SleepingOwl\Admin\Display\Column\Email;
 use SleepingOwl\Admin\Display\Column\Control;
 use SleepingOwl\Admin\Contracts\WithRoutesInterface;
 use SleepingOwl\Admin\Contracts\Display\ColumnInterface;
@@ -62,15 +59,6 @@ class DisplayDatatablesAsync extends DisplayDatatables implements WithRoutesInte
      * @var
      */
     protected $displayLength = false;
-
-    /**
-     * @var array
-     */
-    protected $searchableColumns = [
-        Text::class,
-        Link::class,
-        Email::class,
-    ];
 
     /**
      * DisplayDatatablesAsync constructor.
@@ -241,23 +229,6 @@ class DisplayDatatablesAsync extends DisplayDatatables implements WithRoutesInte
     }
 
     /**
-     * Check if column is searchable.
-     *
-     * @param ColumnInterface $column
-     * @return bool
-     */
-    public function checkSearchableColumns(ColumnInterface $column)
-    {
-        foreach ($this->searchableColumns as $searchableColumn) {
-            if ($column instanceof $searchableColumn) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
      * Apply search to the query.
      *
      * @param Builder $query
@@ -274,7 +245,7 @@ class DisplayDatatablesAsync extends DisplayDatatables implements WithRoutesInte
             $columns = $this->getColumns()->all();
 
             foreach ($columns as $column) {
-                if ($this->checkSearchableColumns($column)) {
+                if ($column->isSearchable()) {
                     if ($column instanceof ColumnInterface) {
                         if (($metaInstance = $column->getMetaData()) instanceof ColumnMetaInterface) {
                             if (method_exists($metaInstance, 'onSearch')) {
@@ -339,8 +310,7 @@ class DisplayDatatablesAsync extends DisplayDatatables implements WithRoutesInte
     }
 
     /**
-     * @return Collection
-     * @throws \Exception
+     * @return void
      */
     public function getCollection()
     {