+++ src/Facades/FormElement.php
@@ -33,6 +33,8 @@ use Illuminate\Database\Eloquent\Model;
  * @method static \SleepingOwl\Admin\Form\Element\Wysiwyg wysiwyg($name, $label = null, $editor = null)
  * @method static \SleepingOwl\Admin\Form\Element\Upload upload($name, $label = null)
  * @method static \SleepingOwl\Admin\Form\Element\Number number($name, $label = null)
+ * @method static \SleepingOwl\Admin\Form\Related\Forms\HasMany hasMany(string $relationName, array $elements)
+ * @method static \SleepingOwl\Admin\Form\Related\Forms\ManyToMany manyToMany(string $relationName, array $elements)
  */
 class FormElement extends Facade
 {